RiderTimes
Successfully search for a specific rider using their Rider ID.
Using datetimepicker instead of textbox and changing its format to time.
Disabling the radio buttons and datetimepicker at the start and enabling specific legs that the rider has entered in when a rider is searched. Also the datetimepickers are enabled for the user to enter the rider's start and end time.